ex6: parameter panic
i realized that this.upKey and this.downKey would be more logical to use in paddle.js, and was subsequently confused when it didnt work. then i took a look at the parameters in script.js that were set up for each paddle and saw that they were either in the wrong order, or some were missing. i fixed that.