P3: FINISHED (see description)
- added final object dialogue
- commented code

im using my home computer which lags a LOT, so i dont know if the frame rate dropped because of the programming or if my computer is a dumpster fire.

im also aware of a bug that usually only affects collisions with ereth where the friendsTalkedTo variable won't pick up her collision. i have NO IDEA why, but if you talk to ereth first it doesn't usually happen and the game functions normally.